# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Тестовое задние на позицию Golang junior в ЦАРКА",
        "title": "TSARKA test tasks API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/rest",
    "paths": {
        "/email/check": {
            "post": {
                "description": "Находит все emails по паттерну  «Email:__$email»",
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "text/plain"
                ],
                "summary": "find emails",
                "parameters": [
                    {
                        "description": "Input string",
                        "name": "str",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/substr/find": {
            "post": {
                "description": "Находит самую длинную подстроку без повторений",
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "text/plain"
                ],
                "summary": "find substring",
                "parameters": [
                    {
                        "description": "Input string",
                        "name": "str",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}